* Fixed bug in getCoord (int converted into qreal) * Sound volume comparisons now are compared with rint() function * Fixed ATH algorithm: if button was pressed before MAX_VOLUME, no max * volume will be played

+++ audiogramchart.h
@@ -76,7 +76,7 @@ private:
     void paintLeft(QPainter *painter);
     void paintRight(QPainter *painter);
 
-    QPoint getCoords(int frequency, int decibel);
+    QPoint getCoords(qreal frequency, qreal decibel);
 };
 
 class AudiogramChartWidget : public QQuickImageProvider
